Add new testcase with all scenarios for segments
MAILPOET-3047
